Moves the OrderedDictTrait to node utils and uses it in coordinates. The OrderedDictTrait now just aliases the Dict trait for python 3.6 and above.
